new component fieldWrapper that splits the code for better readability
